fixed some error with hitbox design

+++ Data/DamageData.cs
@@ -4,6 +4,6 @@ using Godot;
 
 namespace GameSystem.Data.Instance;
     public abstract partial class DamageData : Node{
-        [Export] public double Value { get; set; }
+        [Export] public float Value { get; set; }
         public List<Effect> EffectsValue { get; set; } = new();
         }
